move api js classes to separate folder
